003: the last four useless casts, and one that was not
`decast.py` and `uncast.py` are two of the eight tools `proven.sh` lists as
never having reported anything, and their zeros turn out to be right: all four
of the compiler's remaining `-Wuseless-cast` warnings are in `bmf.cpp`, which
neither tool scans.

Three of the four were real.  `(int32_t)(x - y)` on two `int32_t`s in
`__OFSUB__`, and `~(uintptr_t)(BMF_PAGE - 1)` where the other operand already
carries the type.

The fourth was not, and taking the warning at its word moved `x_ep` by four
bytes.  `__PAIR64__(high, low)` casts `low` to `uint32_t` before the OR; the
compiler is right that at two of the three expansions `low` arrives as a
`uint32_t` already, and wrong that the cast is therefore removable, because the
macro is textual and the third expansion passes a signed value that
sign-extends through the OR and fills the high half with ones.

The fix is to stop the two innocent sites from casting `low` themselves -- they
pass a `uint8_t *`, so the macro's cast does that conversion instead of
repeating one -- which leaves the macro correct and no expansion useless.

A ratchet is a scoreboard and not an instruction: 1345 to 1341, with the
warning that was wrong written down beside the code it was wrong about.

+++ 003/bmf.cpp
@@ -118,6 +118,12 @@
 
 // Two 32-bit halves as one 64-bit value, which is how Hex-Rays writes a
 // register pair.
+// The cast on `low` is load-bearing: without it a signed `low` sign-extends
+// through the OR and fills the high half with ones.  Removing it on the
+// strength of `-Wuseless-cast` moved `x_ep` by four bytes -- the warning was
+// right about two of the three expansions and the macro is textual, so it has
+// to be correct for all three.  The two call sites that were casting `low`
+// themselves no longer do, which is what made the cast look useless.
 #define __PAIR64__(high, low) (((uint64_t)(high) << 32) | (uint32_t)(low))
 
 // |x| as an unsigned value.  The sign of the *result* is the point: the bodies
@@ -131,7 +137,7 @@ inline uint32_t abs32(int32_t x) { return x >= 0 ? (uint32_t)x : (uint32_t)-x; }
 inline int8_t __OFSUB__(int32_t x, int32_t y)
 {
   int8_t sx = (int8_t)(x < 0);
-  return (int8_t)((sx ^ (int8_t)(y < 0)) & (sx ^ (int8_t)((int32_t)(x - y) < 0)));
+  return (int8_t)((sx ^ (int8_t)(y < 0)) & (sx ^ (int8_t)((x - y) < 0)));
 }
 
 // defs.h stops at SLODWORD/SHIDWORD; the bodies also index DWORD lanes
@@ -281,7 +287,7 @@ static void *bmf_page_alloc(size_t n) {
   if (!base)
     return nullptr;
   uintptr_t raw = (uintptr_t)base + sizeof(void *);
-  char *p = (char *)((raw + (BMF_PAGE - 1)) & ~(uintptr_t)(BMF_PAGE - 1));
+  char *p = (char *)((raw + (BMF_PAGE - 1)) & ~(BMF_PAGE - 1));
   ((void **)p)[-1] = base;
   memset(p, 0, n);
   return p;
